# Improve dest.Calculate() and dest.GetMultiBestPath() performance