--- a/internal/pkg/table/message.go
+++ b/internal/pkg/table/message.go
@@ -352,7 +352,7 @@ func newPackerMP(f bgp.Family) *packerMP {
 
 type packerV4 struct {
 	packer
-	hashmap     map[uint32][]*cage
+	hashmap     map[uint64][]*cage
 	mpPaths     []*Path
 	withdrawals []*Path
 }
@@ -488,7 +488,7 @@ func newPackerV4(f bgp.Family) *packerV4 {
 		packer: packer{
 			family: f,
 		},
-		hashmap:     make(map[uint32][]*cage),
+		hashmap:     make(map[uint64][]*cage),
 		withdrawals: make([]*Path, 0),
 		mpPaths:     make([]*Path, 0),
 	}