--- a/internal/pkg/table/path.go
+++ b/internal/pkg/table/path.go
@@ -29,6 +29,7 @@ import (
 	"github.com/osrg/gobgp/v4/pkg/config/oc"
 	"github.com/osrg/gobgp/v4/pkg/log"
 	"github.com/osrg/gobgp/v4/pkg/packet/bgp"
+	"github.com/segmentio/fasthash/fnv1a"
 )
 
 const (
@@ -72,10 +73,7 @@ func (b *Bitmap) FindandSetZeroBit() (uint, error) {
 }
 
 func (b *Bitmap) Expand() {
-	old := b.bitmap
-	new := make([]uint64, len(old)+1)
-	copy(new, old)
-	b.bitmap = new
+	b.bitmap = append(b.bitmap, uint64(0))
 }
 
 func NewBitmap(size int) *Bitmap {
@@ -137,7 +135,7 @@ type Path struct {
 	parent    *Path
 	pathAttrs []bgp.PathAttributeInterface
 	dels      []bgp.BGPAttrType
-	attrsHash uint32
+	attrsHash uint64
 	rejected  bool
 	// doesn't exist in the adj
 	dropped bool
@@ -180,7 +178,6 @@ func NewPath(source *PeerInfo, nlri bgp.AddrPrefixInterface, isWithdraw bool, pa
 	if !isWithdraw && pattrs == nil {
 		return nil
 	}
-
 	return &Path{
 		info: &originInfo{
 			nlri:               nlri,
@@ -553,6 +550,7 @@ func (path *Path) getPathAttr(typ bgp.BGPAttrType) bgp.PathAttributeInterface {
 }
 
 func (path *Path) setPathAttr(a bgp.PathAttributeInterface) {
+	path.attrsHash = 0
 	if len(path.pathAttrs) == 0 {
 		path.pathAttrs = []bgp.PathAttributeInterface{a}
 	} else {
@@ -567,6 +565,7 @@ func (path *Path) setPathAttr(a bgp.PathAttributeInterface) {
 }
 
 func (path *Path) delPathAttr(typ bgp.BGPAttrType) {
+	path.attrsHash = 0
 	if len(path.dels) == 0 {
 		path.dels = []bgp.BGPAttrType{typ}
 	} else {
@@ -1034,19 +1033,44 @@ func (lhs *Path) Equal(rhs *Path) bool {
 		return false
 	}
 
-	if !lhs.GetSource().Equal(rhs.GetSource()) {
+	lhsPathAttrs := lhs.GetPathAttrs()
+	rhsPathAttrs := rhs.GetPathAttrs()
+	// comparing by length first as it's quite fast (with golang slice)
+	// and easy to know if paths are different
+	if len(lhsPathAttrs) == 0 && len(rhsPathAttrs) == 0 {
+		return lhs.EqualBySourceAndPathID(rhs)
+	}
+
+	if len(lhsPathAttrs) != len(rhsPathAttrs) {
 		return false
 	}
 
-	pattrs := func(arg []bgp.PathAttributeInterface) []byte {
-		ret := make([]byte, 0)
-		for _, a := range arg {
-			aa, _ := a.Serialize()
-			ret = append(ret, aa...)
-		}
-		return ret
+	if !lhs.GetSource().Equal(rhs.GetSource()) {
+		return false
 	}
-	return bytes.Equal(pattrs(lhs.GetPathAttrs()), pattrs(rhs.GetPathAttrs()))
+	// The idea here is to calculate the hash of the attributes on demand
+	if lhs.attrsHash > 0 && rhs.attrsHash > 0 { // direct access to the hash to avoid unnecessary hash calculation
+		return lhs.attrsHash == rhs.attrsHash
+	}
+	// slow path comparison, could happen as attributes flags is not part of the hash
+	for t, a := range lhsPathAttrs {
+		b := rhsPathAttrs[t]
+		if a.GetType() != b.GetType() {
+			return false
+		}
+		if a.Len() != b.Len() {
+			return false
+		}
+		if a.GetFlags() != b.GetFlags() {
+			return false
+		}
+	}
+	// really slow path comparison, if hash not been calculated yet
+	if lhs.GetHash() != rhs.GetHash() {
+		return false
+	}
+
+	return true
 }
 
 func (path *Path) MarshalJSON() ([]byte, error) {
@@ -1286,11 +1310,23 @@ func (p *Path) ToLocal() *Path {
 	return path
 }
 
-func (p *Path) SetHash(v uint32) {
+func (p *Path) updateHash() {
+	p.attrsHash = fnv1a.Init64
+	for _, a := range p.GetPathAttrs() {
+		d, _ := a.Serialize()
+		p.attrsHash = fnv1a.AddBytes64(p.attrsHash, d)
+	}
+}
+
+func (p *Path) SetHash(v uint64) {
 	p.attrsHash = v
 }
 
-func (p *Path) GetHash() uint32 {
+// GetHash returns the hash value of the path attributes.
+func (p *Path) GetHash() uint64 {
+	if p.attrsHash == 0 {
+		p.updateHash()
+	}
 	return p.attrsHash
 }