Improve performance by 10% ish by caching path attributes via a path attributes hash if needed

```
Before
BenchmarkMultiPath/Benchmark_Calculate-12                 598702              1837 ns/op             376 B/op         14 allocs/op
BenchmarkMultiPath/Benchmark_GetMultiBestPath-12         3420961               343.5 ns/op             0 B/op          0 allocs/op

After
BenchmarkMultiPath/Benchmark_Calculate-12         	  637652	      1764 ns/op	     376 B/op	      14 allocs/op
BenchmarkMultiPath/Benchmark_GetMultiBestPath-12  	 4771671	       298.0 ns/op	       0 B/op	       0 allocs/op