@nplanelcisco 2025-07-22
> The zebra-nht failure doesn't look like a fluke?

Seems a regression, I will take a look tomorrow