@fujita 2025-07-25
> > Thank you. Just out of curiosity, what caused the zebra tests to fail?
> 
> I forgot to invalidate the `path.hashAttr` if a `add/delAttr` is called. What cause the issue is that Metrics (MED) was wrong for the path, as value changed but not the hash, so compare would return the wrong path.

Understood.
 
> I will add some tests for corner case like this and other behavior check (GetListAttrs() in order, ...)

Great, there should be tests other than the zebra tests that can catch that bug. Thanks!